Create fields for grounded checking in Gem class
The Transform to be used for grounded checking is needed so that I can
easily position where the grounded checking raycast will start.

Note that you cannot simply use transform.position in Physics2D.Raycast
since the raycast will detect the collider if the raycast's starting
position is in a collider. I didn't want this to happen.