Create fallDestination field in Gem class
This is the position the gem GameObject should be at when it finishes
falling. This will be used to check if the code that handles falling
causes the gem to meet or exceed the given destination. When this
happens, the gem's position will be set to the value in this field.