Update gem generation positioning
Instead of spawning a gem a couple of units above its current computed
position (i.e. the position the gem should be based on its row and
column values), gems are now spawned at a fixed height above the board.
An offset will be applied to the height based on which row the gem will
go to (e.g. if the gem belongs to a column that is almost empty and it
is generated for the very top row, then that gem will have a very high
height offset, allowing it to come in last when gems are falling).