Update logs EnsureNoMatches method
Changed LogAsErrors call to Log since not all failed removals are
erroneous. An example situation will be the board having a different
set of gem types used in gem generation from the gem types present
on the board
[e.g. board only generates red and orange gems but the board currently
has yellow and green gems, which causes List.Remove() to return a false
since yellow and green aren't types present in the list of applicable
gem types (which will match with the types used in gem generation)].

Also added "Horizontal matching: " and "Vertical matching: " prefixes
which help to differentiate whether the log came from the horizontal
matches detection portion or the vertical matches detection portion
of the code.