Transition scripts to use the new GemBoard class
A lot have changed. Several methods have the parameter type updated from
GemBehaviour[,] to GemBoard. In some cases, I explicitly used the
GemBehaviour property on the Gem class since it wasn't possible for the
code to compile (e.g. explicitly using the property in if
statements that rely on UnityEngine.Object giving a Boolean) or it isn't
worthwhile creating a property to redirect access (e.g. I explicitly
used the GemBehaviour propety when accessing fallDestination and
isFalling since these are use cases that don't appear often in the
code).