Update board behaviour to use Gem row/col props
Updated GemBoardBehaviour to use the new RowOnBoard and ColOnBoard
properties in the Gem class.

Note that this is not a complete transformation because the GemBehaviour
script contains another set of fields tracking its current row and
column of the gem on the board.
Those fields will need to be removed as part of my effort to isolate the
GemBehaviour script from the Gem class.